requirements.txt
streamlit
pandas
numpy
openai
PyPDF2
pdfminer.six
sentence-transformers
torch
xlsxwriter
bcrypt
python-dotenv
faiss-cpu
transformers
python-docx
openpyxl==3.1.5
email-validator==2.2.0